feat: adiciona funcionalidade de geração de faturas em PDF e rotas correspondentes; implementa exclusão de imagens de produtos

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file')
 
 const Product = require('../models/product');
 
@@ -138,6 +139,9 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if(image){
+        if(product.imageUrl){
+          fileHelper.deleteFile(product.imageUrl);
+        }
         product.imageUrl = image.path;
       }
 
@@ -173,15 +177,30 @@ exports.getProducts = (req, res, next) => {
     });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      res.redirect('/admin/products');
-    })
-    .catch(err => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
-    });
+exports.postDeleteProduct = async (req, res, next) => {
+  try{
+    const prodId = req.body.productId;
+    const product = await Product.findById(prodId);
+
+    if(!product){
+      return next(new Error('Product Not Found'))
+    }
+
+    if(product.imageUrl){
+      fileHelper.deleteFile(product.imageUrl);
+    }
+    
+    Product.deleteOne({ _id: prodId, userId: req.user._id })
+      .then(() => {
+        res.redirect('/admin/products');
+      })
+      .catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+      });
+  }catch(err){
+    next(err)
+  }
+  
 };
